+++ packages/selector-resolve-nested/test/flatten-nested-selector/parent.mjs
@@ -1,13 +1,13 @@
 import assert from 'node:assert/strict';
 import test from 'node:test';
-import { desugarNestedSelector } from '@csstools/selector-resolve-nested';
+import { flattenNestedSelector } from '@csstools/selector-resolve-nested';
 import parser from 'postcss-selector-parser';
 
 test('parent', async () => {
 	const a = parser().astSync('.a .aa');
 	const b = parser().astSync(':is(&) + .b');
 
-	const ba = desugarNestedSelector(b, a);
+	const ba = flattenNestedSelector(b, a);
 
 	const firstNode = ba.nodes[0].first;
 	assert.equal(firstNode.type, 'pseudo');