+++ packages/selector-resolve-nested/test/index.mjs
@@ -7,8 +7,8 @@ import './resolve-nested-selector/parent.mjs';
 import './resolve-nested-selector/simple-compound-complex.mjs';
 import './resolve-nested-selector/source-indices.mjs';
 
-import './desugar-nested-selector/lists.mjs';
-import './desugar-nested-selector/multiple-resolves.mjs';
-import './desugar-nested-selector/parent.mjs';
-import './desugar-nested-selector/simple-compound-complex.mjs';
-import './desugar-nested-selector/source-indices.mjs';
+import './flatten-nested-selector/lists.mjs';
+import './flatten-nested-selector/multiple-resolves.mjs';
+import './flatten-nested-selector/parent.mjs';
+import './flatten-nested-selector/simple-compound-complex.mjs';
+import './flatten-nested-selector/source-indices.mjs';